requirements.txt
h5py
numpy
scipy
ipykernel
matplotlibtqdm